+++ src/Switch/test.js
@@ -1,0 +1,24 @@
+import React from 'react'
+import Switch from './'
+import renderer from 'react-test-renderer'
+
+
+it('Default Switch', () => {
+  const tree = renderer
+    .create(
+      <Switch>Label text</Switch>
+    )
+    .toJSON()
+
+  expect(tree).toMatchSnapshot()
+})
+
+it('Extra class Switch', () => {
+  const tree = renderer
+    .create(
+      <Switch extraClass="foo">Label text</Switch>
+    )
+    .toJSON()
+
+  expect(tree).toMatchSnapshot()
+})