+++ src/stories.js
@@ -1,2 +1,2 @@
 // STYLES
-import '../dist/index.css'
+import './index.css'