+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Component `Switch` use `children` to define the label content, instead `label` and `htmlLabel` properties. [#28]
+- The signature of `TableCell` component has changed. See [#37] and [#65] for more info.
+
 ### Fixed
 - Removed `display:flex` to Check areas.
 - Decrease padding in Check areas in columns.
@@ -165,12 +169,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#20]: https://github.com/marketgoo/Ola/issues/20
 [#21]: https://github.com/marketgoo/Ola/issues/21
 [#23]: https://github.com/marketgoo/Ola/issues/23
+[#28]: https://github.com/marketgoo/Ola/issues/28
 [#30]: https://github.com/marketgoo/Ola/issues/30
 [#32]: https://github.com/marketgoo/Ola/issues/32
 [#33]: https://github.com/marketgoo/Ola/issues/33
 [#34]: https://github.com/marketgoo/Ola/issues/34
 [#35]: https://github.com/marketgoo/Ola/issues/35
 [#36]: https://github.com/marketgoo/Ola/issues/36
+[#37]: https://github.com/marketgoo/Ola/issues/37
 [#39]: https://github.com/marketgoo/Ola/issues/39
 [#40]: https://github.com/marketgoo/Ola/issues/40
 [#41]: https://github.com/marketgoo/Ola/issues/41
@@ -179,6 +185,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#47]: https://github.com/marketgoo/Ola/issues/47
 [#54]: https://github.com/marketgoo/Ola/issues/54
 [#56]: https://github.com/marketgoo/Ola/issues/56
+[#65]: https://github.com/marketgoo/Ola/issues/65
 
 [Unreleased]: https://github.com/marketgoo/Ola/compare/v0.3.0...HEAD
 [0.3.0]: https://github.com/marketgoo/Ola/compare/v0.2.2...v0.3.0