+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.4.0] - Unreleased
+### Added
+- Property `extraClass` to `Switch`, `ModalHeader`, `PanelHeader`.
+
 ### Changed
-- Component `Switch` use `children` to define the label content, instead `label` and `htmlLabel` properties. [#28]
+- `Switch`: From now use `children` to define the label content, instead `label` and `htmlLabel` properties. [#28]
+- `ProgressBar`: From now use `children` to define the description, instead `description` and `htmlDescription` properties. [#28]
 - The signature of `TableCell` component has changed. See [#37] and [#65] for more info.
-- Removed the properties `htmlTitle` and `htmlIntro` in `ModalHeader`. Now `title` and `intro` accept not only strings but also nodes as values [#28]
+- `ModalHeader`: Removed the properties `htmlTitle` and `htmlIntro`. From now, `title` and `intro` accept strings and nodes as values [#28]
+- `PanelHeader`: Removed the properties `htmlTitle` and `htmlIntro`. From now, `title` and `intro` accept strings and nodes as values [#28]
+- `PanelContent`: Removed the property `htmlTitle`. From now `title` accepts strings and nodes as values [#28]
+- `Metric`: Removed the property `htmlFooter`. From now `footer` accepts strings and nodes as values [#28]
+- `Task`: Removed the property `htmlTitle`. From now `title` accepts strings and nodes as values [#28]
 
 ### Fixed
 - Removed `display:flex` to Check areas.
@@ -188,7 +196,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#56]: https://github.com/marketgoo/Ola/issues/56
 [#65]: https://github.com/marketgoo/Ola/issues/65
 
-[Unreleased]: https://github.com/marketgoo/Ola/compare/v0.3.0...HEAD
+[0.4.0]: https://github.com/marketgoo/Ola/compare/v0.3.0...HEAD
 [0.3.0]: https://github.com/marketgoo/Ola/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/marketgoo/Ola/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/marketgoo/Ola/compare/v0.2.0...v0.2.1